+++ src/pages/NySykmelding/components/skjemaComponents/AvbrytDialog.less
@@ -1,0 +1,3 @@
+.avbrytdialog--margin-bottom {
+    margin-bottom: 1rem;
+}
